+++ libs/shared/tests/lib/modules/invoices/repos/implementations/knexInvoiceItemRepo.steps.ts
@@ -1,6 +1,10 @@
 import { expect } from 'chai';
 import { Given, When, Then, Before } from '@cucumber/cucumber';
 
+import { GuardFailure } from '../../../../../../src/lib/core/logic/GuardFailure';
+import { RepoError } from '../../../../../../src/lib/infrastructure/RepoError';
+import { Either } from '../../../../../../src/lib/core/logic/Either';
+
 import { InvoiceItemId } from '../../../../../../src/lib/modules/invoices/domain/InvoiceItemId';
 import { InvoiceItemMap } from '../../../../../../src/lib/modules/invoices/mappers/InvoiceItemMap';
 import { InvoiceItem } from '../../../../../../src/lib/modules/invoices/domain/InvoiceItem';
@@ -9,13 +13,19 @@ import { UniqueEntityID } from './../../../../../../src/lib/core/domain/UniqueEn
 import { MockInvoiceItemRepo } from './../../../../../../src/lib/modules/invoices/repos/mocks/mockInvoiceItemRepo';
 
 function makeInvoiceItemData(overwrites?: any): InvoiceItem {
-  return InvoiceItemMap.toDomain({
+  const invoiceItem = InvoiceItemMap.toDomain({
     manuscriptId: 'manuscript-id-1',
     invoiceId: 'invoice-id-1',
     price: 0,
     dateCreated: new Date(),
     ...overwrites,
   });
+
+  if (invoiceItem.isLeft()) {
+    throw invoiceItem.value;
+  }
+
+  return invoiceItem.value;
 }
 
 let mockInvoiceItemRepo: MockInvoiceItemRepo;
@@ -23,9 +33,12 @@ let invoiceItem: InvoiceItem;
 let saveInvoiceItem: InvoiceItem;
 let foundInvoiceItem: InvoiceItem;
 let invoiceItemExists: boolean;
+let maybeFoundInvoiceItem: Either<GuardFailure | RepoError, InvoiceItem>;
 
-Before(async () => {
+Before({ tags: '@ValidateKnexInvoiceItemRepo' }, async () => {
   mockInvoiceItemRepo = new MockInvoiceItemRepo();
+  maybeFoundInvoiceItem = null;
+  invoiceItem = null;
 });
 
 Given(
@@ -42,9 +55,15 @@ When(
     const invoiceItemIdObj = InvoiceItemId.create(
       new UniqueEntityID(invoiceItemId)
     );
-    foundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
+    const maybeFoundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
       invoiceItemIdObj
     );
+
+    if (maybeFoundInvoiceItem.isLeft()) {
+      throw maybeFoundInvoiceItem.value;
+    }
+
+    foundInvoiceItem = maybeFoundInvoiceItem.value;
   }
 );
 
@@ -56,7 +75,13 @@ When(
   /^we call getInvoiceItemById for an un-existent invoice item "([\w-]+)"$/,
   async (wrongInvoiceItemId: string) => {
     const id = InvoiceItemId.create(new UniqueEntityID(wrongInvoiceItemId));
-    foundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(id);
+    maybeFoundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(id);
+
+    if (maybeFoundInvoiceItem.isRight()) {
+      foundInvoiceItem = maybeFoundInvoiceItem.value;
+    } else {
+      foundInvoiceItem = null;
+    }
   }
 );
 
@@ -70,10 +95,21 @@ When(
     const invoiceItemIdObj = InvoiceItemId.create(
       new UniqueEntityID(invoiceItemId)
     );
-    foundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
+    const maybeFoundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
       invoiceItemIdObj
     );
-    await mockInvoiceItemRepo.delete(foundInvoiceItem);
+
+    if (maybeFoundInvoiceItem.isLeft()) {
+      throw maybeFoundInvoiceItem.value;
+    }
+
+    foundInvoiceItem = maybeFoundInvoiceItem.value;
+
+    const maybeResult = await mockInvoiceItemRepo.delete(foundInvoiceItem);
+
+    if (maybeResult.isLeft()) {
+      throw maybeResult.value;
+    }
   }
 );
 
@@ -97,12 +133,24 @@ When(
     const invoiceItemIdObj = InvoiceItemId.create(
       new UniqueEntityID(invoiceItemId)
     );
-    foundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
+    const maybeFoundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
       invoiceItemIdObj
     );
+
+    if (maybeFoundInvoiceItem.isLeft()) {
+      throw maybeFoundInvoiceItem.value;
+    }
+
+    foundInvoiceItem = maybeFoundInvoiceItem.value;
+
     const invoiceItemPrice = 0;
     foundInvoiceItem.price = invoiceItemPrice;
-    await mockInvoiceItemRepo.update(foundInvoiceItem);
+
+    const maybeResult = await mockInvoiceItemRepo.update(foundInvoiceItem);
+
+    if (maybeResult.isLeft()) {
+      throw maybeResult.value;
+    }
   }
 );
 
@@ -112,9 +160,16 @@ Then(
     const invoiceItemIdObj = InvoiceItemId.create(
       new UniqueEntityID(invoiceItemId)
     );
-    foundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
+    const maybeFoundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(
       invoiceItemIdObj
     );
+
+    if (maybeFoundInvoiceItem.isLeft()) {
+      throw maybeFoundInvoiceItem.value;
+    }
+
+    foundInvoiceItem = maybeFoundInvoiceItem.value;
+
     expect(foundInvoiceItem.price).to.equal(0);
   }
 );
@@ -123,11 +178,28 @@ When(
   /^we call exists for ([\w-]+) invoice item id$/,
   async (invoiceItemId: string) => {
     const id = InvoiceItemId.create(new UniqueEntityID(invoiceItemId));
-    foundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(id);
-    if (!foundInvoiceItem) {
-      foundInvoiceItem = await makeInvoiceItemData({ id: invoiceItemId });
+    maybeFoundInvoiceItem = await mockInvoiceItemRepo.getInvoiceItemById(id);
+
+    if (maybeFoundInvoiceItem.isRight()) {
+      foundInvoiceItem = maybeFoundInvoiceItem.value;
+
+      if (!foundInvoiceItem) {
+        foundInvoiceItem = await makeInvoiceItemData({ id: invoiceItemId });
+      }
+
+      const maybeInvoiceItemExists = await mockInvoiceItemRepo.exists(
+        foundInvoiceItem
+      );
+
+      if (maybeInvoiceItemExists.isLeft()) {
+        throw maybeInvoiceItemExists.value;
+      }
+
+      invoiceItemExists = maybeInvoiceItemExists.value;
+    } else {
+      invoiceItem = null;
+      invoiceItemExists = false;
     }
-    invoiceItemExists = await mockInvoiceItemRepo.exists(foundInvoiceItem);
   }
 );
 
@@ -143,7 +215,13 @@ Given(
 );
 
 When('we call InvoiceItem.save on the invoice item object', async () => {
-  saveInvoiceItem = await mockInvoiceItemRepo.save(invoiceItem);
+  const maybeSaveInvoiceItem = await mockInvoiceItemRepo.save(invoiceItem);
+
+  if (maybeSaveInvoiceItem.isLeft()) {
+    throw maybeSaveInvoiceItem.value;
+  }
+
+  saveInvoiceItem = maybeSaveInvoiceItem.value;
 });
 
 Then('the invoice item object should be saved', async () => {
